+++ lab/src/main/java/Main.java
@@ -2,6 +2,6 @@ public class Main {
 
     public static void main(String[] args) {
         DES des = new DES();
-        int[][] subKeys = des.keyToSubKeys("IEOFIT#1");
+        byte[][] subKeys = des.keyToSubKeys("IEOFIT#1");
     }
 }
